fix(core): enforce nanosecond precision for OCC and secure daemon teardown
This PR wraps up the v1.9.10 Concurrency milestone.
- Closes #38: Replaced all float-based st_mtime with st_mtime_ns for exact OCC and cache validation. Added normalize_stored_mtime_ns for backward compatibility.
- Closes #44: Wrapped terminal state flushes in the daemon's exit path with strict exception logging to prevent silent teardown failures.

+++ src/agent/plumber_modules/entity_consolidation.py
@@ -47,7 +47,7 @@ def run_entity_consolidation(
         if should_skip_entity_overlap_pair(graph_root, page_title, target):
             continue
 
-        baselines: dict[str, float | None] = {}
+        baselines: dict[str, int | None] = {}
         for candidate in (page_title, target):
             try:
                 candidate_path = graph_safe_page_path(graph_root, candidate)
